+++ application/main-application/src/main/java/org/mainapplication/global/constants/HeaderConstants.java
@@ -1,0 +1,11 @@
+package org.mainapplication.global.constants;
+
+public final class HeaderConstants {
+	private HeaderConstants() {
+		throw new UnsupportedOperationException("인스턴트를 생성할 수 없습니다.");
+	}
+	public static final String TOKEN_PREFIX = "Bearer ";
+	public static final String ACCESS_TOKEN_HEADER = "Authorization";
+	public static final String REFRESH_TOKEN_HEADER = "RefreshToken";
+
+}
