+++ domain/domain-module/build.gradle
@@ -7,7 +7,6 @@ plugins {
 version = '0.0.1-SNAPSHOT'
 
 dependencies {
-	api project(':common:common-module')
 	api 'org.springframework.boot:spring-boot-starter-data-jpa'
 
 	// db
