--- domain/domain-module/src/main/java/org/domainmodule/user/service/UserService.java
@@ -1,9 +1,0 @@
-package org.domainmodule.user.service;
-
-import org.commonmodule.dto.OAuth2Response;
-import org.domainmodule.user.entity.User;
-
-public interface UserService {
-	User loginOrCreateUser(OAuth2Response oAuth2Response);
-	User registerUser(OAuth2Response oAuth2Response);
-}
