--- inner-system/security/build.gradle
@@ -1,28 +1,0 @@
-plugins {
-    id 'java'
-    id 'org.springframework.boot' version '3.4.1'
-    id 'io.spring.dependency-management' version '1.1.7'
-}
-
-version = '0.0.1-SNAPSHOT'
-
-dependencies {
-
-    api project(':domain:domain-module')
-
-    // Security
-    api 'org.springframework.boot:spring-boot-starter-security'
-    testImplementation 'org.springframework.security:spring-security-test'
-
-    //Oauth2
-    api 'org.springframework.boot:spring-boot-starter-oauth2-client'
-
-    //jwt
-    api 'io.jsonwebtoken:jjwt-api:0.11.5'
-    api 'io.jsonwebtoken:jjwt-impl:0.11.5'
-    api 'io.jsonwebtoken:jjwt-jackson:0.11.5'
-}
-
-tasks.named('test') {
-    useJUnitPlatform()
-}
