--- inner-system/security/src/main/java/org/security/SecurityApplication.java
@@ -1,15 +1,0 @@
-package org.security;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication
-@ComponentScan(basePackages = {"org.security", "org.domainmodule"})
-public class SecurityApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SecurityApplication.class, args);
-	}
-
-}
