--- inner-system/security/src/main/java/org/security/entity/CustomUserDetails.java
@@ -1,37 +1,0 @@
-package org.security.entity;
-
-import java.util.Collection;
-import java.util.Map;
-
-import org.domainmodule.user.entity.User;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class CustomUserDetails implements OAuth2User {
-
-	private final User user;
-	private Map<String, Object> attributes;
-
-	@Override
-	public Map<String, Object> getAttributes() {
-		return attributes;
-	}
-
-
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		return null;
-	}
-
-	@Override
-	public String getName() {
-		return user.getName();
-	}
-
-	public String getId() {return user.getId().toString();}
-}
