--- inner-system/security/src/main/java/org/security/oauth2/CustomOauth2UserService.java
@@ -1,43 +1,0 @@
-package org.security.oauth2;
-
-import java.util.Map;
-
-import org.commonmodule.dto.OAuth2Response;
-import org.domainmodule.user.entity.User;
-import org.domainmodule.user.service.UserService;
-import org.security.dto.GoogleResponse;
-import org.security.entity.CustomUserDetails;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.stereotype.Service;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class CustomOauth2UserService extends DefaultOAuth2UserService {
-
-	private final UserService userService;
-
-	@Override
-	public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
-		OAuth2User oAuth2User = super.loadUser(userRequest);
-
-		String registrationId = userRequest.getClientRegistration().getRegistrationId();
-		OAuth2Response oAuth2Response = getOAuth2Response(oAuth2User, registrationId);
-		User user = userService.loginOrCreateUser(oAuth2Response);
-		return new CustomUserDetails(user);
-	}
-
-	private OAuth2Response getOAuth2Response(OAuth2User oAuth2User, String registrationId) {
-		Map<String, Object> attributes = oAuth2User.getAttributes();
-		switch (registrationId) {
-			case "google":
-				return GoogleResponse.fromAttributes(attributes);
-			default:
-				throw new OAuth2AuthenticationException("지원되지 않는 로그인 제공자: " + registrationId);
-		}
-	}
-}
