--- inner-system/security/src/main/resources/application.yml
@@ -1,28 +1,0 @@
-spring:
-  security:
-    oauth2:
-      client:
-        registration:
-          google:
-            client-id: ${GOOGLE_CLIENT_ID}
-            client-secret: ${GOOGLE_CLIENT_SECRET}
-            redirect_uri: ${GOOGLE_REDIRECT_URI}
-            authorization-grant-type: authorization_code
-            scope:
-              - email
-              - profile
-        provider:
-          google:
-            authorization-uri: https://accounts.google.com/o/oauth2/auth
-            token-uri: https://oauth2.googleapis.com/token
-            user-info-uri: https://www.googleapis.com/oauth2/v3/userinfo
-jwt:
-  access-token-key: ${ACCESS_JWT_KEY:}
-  refresh-token-key: ${REFRESH_JWT_KEY:}
-  access-token-expiration: ${ACCESS_JWT_EXPIRATION:900000}  # 15분
-  refresh-token-expiration: ${REFRESH_JWT_EXPIRATION:604800000} # 7일
-logging:
-  level:
-    org:
-      springframework:
-        security=DEBUG:
