--- inner-system/security/src/test/java/org/security/SecurityApplicationTests.java
@@ -1,13 +1,0 @@
-package org.security;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SecurityApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
